+++ Sources.cmake
@@ -23,9 +23,9 @@ set(LIBPNG_VERSION 1.6.37)
 set(LIBPNG_URL https://downloads.sourceforge.net/project/libpng/libpng16/${LIBPNG_VERSION}/libpng-${LIBPNG_VERSION}.tar.gz)
 set(LIBPNG_MD5 6c7519f6c75939efa0ed3053197abd54)
 
-set(PYTHON_CMAKE_HASH c8ebbcde77433daf17c6e91f8bc7041c380bc61c)
-set(PYTHON_CMAKE_URL https://github.com/freeorion/python-cmake-buildsystem/archive/${PYTHON_CMAKE_HASH}.zip)
-set(PYTHON_CMAKE_MD5 ef568d9a803ef4fdc7dcd0453742f3ed)
+set(PYTHON_CMAKE_HASH 7fb65a846c3264576b9dcfb933b698e166e6b4a0)
+set(PYTHON_CMAKE_URL https://github.com/o01eg/python-cmake-buildsystem/archive/${PYTHON_CMAKE_HASH}.zip)
+set(PYTHON_CMAKE_MD5 b4790d5af93ead53c6447401170c631c)
 
 set(PYTHON3_VERSION 3.9.21)
 set(PYTHON3_URL https://www.python.org/ftp/python/${PYTHON3_VERSION}/Python-${PYTHON3_VERSION}.tgz)